pyproject.toml
[project]
name = "translator-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1"
]
